+++ ParallaxGenLib/src/patchers/PatcherComplexMaterial.cpp
@@ -58,14 +58,31 @@ auto PatcherComplexMaterial::canApply(NiShape &NIFShape) -> bool {
 }
 
 auto PatcherComplexMaterial::shouldApply(nifly::NiShape &NIFShape, std::vector<PatcherMatch> &Matches) -> bool {
-  return shouldApply(getTextureSet(NIFShape), Matches);
+  const auto OldSlots = getTextureSet(NIFShape);
+
+  Matches.clear();
+
+  // Check for CM matches
+  shouldApply(OldSlots, Matches);
+
+  // check if env mapping is already enabled
+  if (getNIF()->GetShader(&NIFShape)->GetShaderType() == BSLSP_ENVMAP) {
+    // Mesh has env mapping, check if slot in old slots is CM
+    if (getPGD()->getTextureType(OldSlots[static_cast<size_t>(NIFUtil::TextureSlots::ENVMASK)]) == NIFUtil::TextureType::COMPLEXMATERIAL) {
+      // mesh is already patched for CM
+      Logger::trace(L"This shape already has ComplexMaterial");
+      PatcherMatch Match;
+      Match.MatchedPath = getNIFPath().wstring();
+      Matches.push_back(Match);
+    }
+  }
+
+  return !Matches.empty();
 }
 
 auto PatcherComplexMaterial::shouldApply(const std::array<std::wstring, NUM_TEXTURE_SLOTS> &OldSlots,
                                          std::vector<PatcherMatch> &Matches) -> bool {
   static const auto CMBaseMap = getPGD()->getTextureMapConst(NIFUtil::TextureSlots::ENVMASK);
-
-  Matches.clear();
 
   // Search prefixes
   const auto SearchPrefixes = NIFUtil::getSearchPrefixes(OldSlots);
@@ -117,6 +134,11 @@ auto PatcherComplexMaterial::applyPatch(NiShape &NIFShape, const PatcherMatch &M
   // Prep
   ShapeDeleted = false;
 
+  if (Match.MatchedPath == getNIFPath().wstring()) {
+    // Do nothing if the match is the same as the current NIF
+    return getTextureSet(NIFShape);
+  }
+
   // Apply shader
   applyShader(NIFShape, NIFModified);
 
@@ -130,6 +152,11 @@ auto PatcherComplexMaterial::applyPatch(NiShape &NIFShape, const PatcherMatch &M
 auto PatcherComplexMaterial::applyPatchSlots(const std::array<std::wstring, NUM_TEXTURE_SLOTS> &OldSlots,
                                              const PatcherMatch &Match) -> std::array<std::wstring, NUM_TEXTURE_SLOTS> {
   array<wstring, NUM_TEXTURE_SLOTS> NewSlots = OldSlots;
+
+  if (Match.MatchedPath == getNIFPath().wstring()) {
+    // Do nothing if the match is the same as the current NIF
+    return NewSlots;
+  }
 
   const auto MatchedPath = Match.MatchedPath;