+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fixed deleted texture maps not actually deleting
 - Added additional help text to mod sort dialog
 - PBR slot commands will have "textures\\" added to the beginning automatically if not already there
+- Meshes that are already patched for complex material can be ignored depending on priority order
 
 ## [0.7.3] - 2024-12-09